requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo5==1.3.7
dnspython==2.6.1
drf-spectacular==0.27.2
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mongoengine==0.28.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
pymongo==3.12.3
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.3
sqlparse==0.2.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
